+++ src/com/intellij/idea/plugin/hybris/project/descriptors/JavaLibraryDescriptor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -27,6 +27,7 @@ data class JavaLibraryDescriptor(
     var libraryFile: File,
     var jarFiles: Set<File> = emptySet(),
     var sourceFiles: List<File> = emptyList(),
+    var sourceJarDirectories: List<File> = emptyList(),
     var exported: Boolean = false,
     var directoryWithClasses: Boolean = false,
     var descriptorType: LibraryDescriptorType = LibraryDescriptorType.UNKNOWN,